Create better view for resume and contact section
Identity crisis! Site is highly casual but wanting to be business casual 😂

CasuallyDev casual host: "I'm in my shorts and the living space is messy (we have a dog), but would you want some water? There's cookies on the table. Make yourself at home"

CasuallyDev business casual host: "We have a prime selection of comfy slippers. We'll start the hearty dinner in thirty minutes. Remote is on the sofa to use the guest Netflix for your enjoyment, and our Wi-fi password is on the Cafe menu. Make yourself at home"